# optimize next_chunk impls for Filter and FilterMap