@thomcc 2023-05-22
Nice improvement. The code is pretty complex, but seems correct and the benchmarks definitely justify it.

@bors r+ rollup=never